Add render for timetables
