+++ templates/timetable_response.j2
@@ -1,0 +1,6 @@
+{% from 'shared_macros.j2' import timetables %}
+
+```{{ timetables(timetable_names) }}
+You can get menu with the command "menu TIMETABLE_NAME"
+OR type "help" to see a list of more commands.
+```
